+++ network-api/networkapi/donate_banner/models.py
@@ -1,18 +1,10 @@
 from django.contrib import admin
 from django.core.validators import RegexValidator
 from django.db import models
-from wagtail import hooks
 from wagtail.admin.panels import FieldPanel, HelpPanel
-from wagtail.admin.ui.tables import UpdatedAtColumn
-from wagtail.contrib.settings.models import BaseSiteSetting, register_setting
 from wagtail.models import PreviewableMixin, TranslatableMixin
-from wagtail.snippets.models import register_snippet
-from wagtail.snippets.views.snippets import SnippetViewSet
 from wagtail_localize.fields import SynchronizedField, TranslatableField
 
-from networkapi.wagtailcustomization.views.snippet_chooser import (
-    DefaultLocaleSnippetChooserViewSet,
-)
 from networkapi.wagtailpages.constants import url_or_query_regex
 
 
@@ -94,45 +86,3 @@ class DonateBanner(TranslatableMixin, PreviewableMixin, models.Model):
         if self.site_donate_banner.exists():
             return "Yes"
         return "No"
-
-
-# Customise DonateBanner Snippet admin listing to show extra columns.
-class DonateBannerViewSet(SnippetViewSet):
-    model = DonateBanner
-    description = "Donate Banner"
-    list_display = ["name", "cta_link", "is_active", UpdatedAtColumn()]
-
-
-register_snippet(DonateBanner, viewset=DonateBannerViewSet)
-
-
-# Customise chooser to only show the default language banners as options.
-# We do not want editors to select the translations as
-# localisation will be handled on the template instead.
-@hooks.register("register_admin_viewset")
-def register_donate_banner_chooser_viewset():
-    return DefaultLocaleSnippetChooserViewSet(
-        "wagtailsnippetchoosers_custom_donatebanner",
-        model=DonateBanner,
-        url_prefix="donate_banner/chooser",
-    )
-
-
-@register_setting(icon="pick")
-class SiteDonateBanner(BaseSiteSetting):
-    select_related = ["active_donate_banner"]
-
-    active_donate_banner = models.ForeignKey(
-        "donate_banner.DonateBanner",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="site_donate_banner",
-    )
-
-    content_panels = [
-        FieldPanel("active_donate_banner"),
-    ]
-
-    class Meta:
-        verbose_name = "Donate Banner"
